reminder and meds pop-ups show up
